mysite/requirements.txt
asgiref==3.7.2
asttokens==2.4.0
backcall==0.2.0
decorator==5.1.1
Django==4.2.3
django-extensions==3.2.3
django-filter==23.3
django-stubs==4.2.4
django-stubs-ext==4.2.2
djangorestframework==3.14.0
executing==1.2.0
ipython==8.15.0
jedi==0.19.0
matplotlib-inline==0.1.6
mypy==1.5.1
mypy-extensions==1.0.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.10.1
types-pytz==2023.3.1.1
types-PyYAML==6.0.12.12
typing_extensions==4.7.1
wcwidth==0.2.7
